Updated namespaces after file moves https://github.com/achobanov/web-application/issues/3
