+++ tests/Web.IntegrationTests/Pipeline/ArticlesPipelineTests.cs
@@ -2,8 +2,8 @@
 {
     using System.Linq;
     using Blog.Application.Articles.Commands;
+    using Blog.Gateways.Persistence.Providers;
     using Blog.Gateways.Web.Areas.Api;
-    using Gateways.Persistence;
     using MyTested.AspNetCore.Mvc;
     using Shouldly;
     using Xunit;
